chore(tests) Include element resizing in inspector padding tests
