+++ editor/src/components/inspector/sections/layout-section/layout-system-subsection/padding-section.spec.browser2.tsx
@@ -1,0 +1,362 @@
+import { act, fireEvent, RenderResult, screen } from '@testing-library/react'
+import * as EP from '../../../../../core/shared/element-path'
+import { expectSingleUndoStep } from '../../../../../utils/utils.test-utils'
+import { getSubduedPaddingControlTestID } from '../../../../canvas/controls/select-mode/subdued-padding-control'
+import {
+  EditorRenderResult,
+  getPrintedUiJsCode,
+  makeTestProjectCodeWithSnippet,
+  renderTestEditorWithCode,
+  TestScenePath,
+} from '../../../../canvas/ui-jsx.test-utils'
+import { selectComponents } from '../../../../editor/actions/action-creators'
+
+async function getControl(
+  controlTestId: string,
+  renderedDOM: RenderResult,
+): Promise<HTMLInputElement> {
+  return (await renderedDOM.findByTestId(controlTestId)) as HTMLInputElement
+}
+
+async function setControlValue(
+  controlTestId: string,
+  newValue: string,
+  renderedDOM: RenderResult,
+): Promise<void> {
+  const control = await getControl(controlTestId, renderedDOM)
+
+  await act(() => {
+    fireEvent.focus(control)
+    fireEvent.change(control, { target: { value: newValue } })
+    fireEvent.blur(control)
+  })
+}
+
+describe('padding controls shorthand', () => {
+  function makeCodeSnippetWithKeyValue(props: { [key: string]: any }): string {
+    const propsStr = Object.keys(props)
+      .map((k) => `${k}: ${JSON.stringify(props[k])},`)
+      .join('\n')
+    return `
+      <div
+        data-uid='aaa'
+      >
+        <div
+          style={{ boxSizing: 'border-box', ${propsStr} }}
+          data-uid='bbb'
+        >
+          <div
+            style={{ width: 100, height: 100 }}
+            data-uid='ccc'
+          />
+        </div>
+      </div>
+  `
+  }
+
+  const tests = [
+    {
+      name: 'without props',
+      startSnippet: makeCodeSnippetWithKeyValue({ width: 100, height: 100 }),
+      control: async (renderResult: EditorRenderResult) => {
+        await expectSingleUndoStep(renderResult, async () => {
+          await setControlValue('padding-V', '20', renderResult.renderedDOM)
+        })
+      },
+      endSnippet: makeCodeSnippetWithKeyValue({ width: 100, height: 140, padding: '20px 0px' }),
+    },
+    {
+      name: 'with shorthand',
+      startSnippet: makeCodeSnippetWithKeyValue({ width: 120, height: 120, padding: 10 }),
+      control: async (renderResult: EditorRenderResult) => {
+        await expectSingleUndoStep(renderResult, async () => {
+          await setControlValue('padding-one', '20', renderResult.renderedDOM)
+        })
+      },
+      endSnippet: makeCodeSnippetWithKeyValue({ width: 140, height: 140, padding: 20 }),
+    },
+    {
+      name: 'with single value (2-values)',
+      startSnippet: makeCodeSnippetWithKeyValue({ width: 110, height: 100, paddingLeft: 10 }),
+      control: async (renderResult: EditorRenderResult) => {
+        await setControlValue('padding-V', '20', renderResult.renderedDOM)
+      },
+      endSnippet: makeCodeSnippetWithKeyValue({
+        width: 110,
+        height: 140,
+        paddingLeft: 10,
+        paddingTop: 20,
+        paddingBottom: 20,
+      }),
+    },
+    {
+      name: 'with single value (1-value)',
+      startSnippet: makeCodeSnippetWithKeyValue({ width: 110, height: 100, paddingLeft: 10 }),
+      before: async (renderResult: EditorRenderResult) => {
+        await act(async () => {
+          fireEvent.click(screen.getByTestId('padding-cycle-mode'))
+          await renderResult.getDispatchFollowUpActionsFinished()
+          fireEvent.click(screen.getByTestId('padding-cycle-mode'))
+          await renderResult.getDispatchFollowUpActionsFinished()
+        })
+      },
+      control: async (renderResult: EditorRenderResult) => {
+        await expectSingleUndoStep(renderResult, async () => {
+          await setControlValue('padding-one', '20', renderResult.renderedDOM)
+        })
+      },
+      endSnippet: makeCodeSnippetWithKeyValue({
+        width: 140,
+        height: 140,
+        paddingLeft: 20,
+        paddingTop: 20,
+        paddingRight: 20,
+        paddingBottom: 20,
+      }),
+    },
+    {
+      name: 'with multiple values (1-value)',
+      startSnippet: makeCodeSnippetWithKeyValue({
+        width: 130,
+        height: 100,
+        paddingLeft: 10,
+        paddingRight: 20,
+      }),
+      before: async (renderResult: EditorRenderResult) => {
+        await act(async () => {
+          fireEvent.click(screen.getByTestId('padding-cycle-mode'))
+          await renderResult.getDispatchFollowUpActionsFinished()
+          fireEvent.click(screen.getByTestId('padding-cycle-mode'))
+          await renderResult.getDispatchFollowUpActionsFinished()
+        })
+      },
+      control: async (renderResult: EditorRenderResult) => {
+        await expectSingleUndoStep(renderResult, async () => {
+          await setControlValue('padding-one', '20', renderResult.renderedDOM)
+        })
+      },
+      endSnippet: makeCodeSnippetWithKeyValue({
+        width: 140,
+        height: 140,
+        paddingLeft: 20,
+        paddingRight: 20,
+        paddingTop: 20,
+        paddingBottom: 20,
+      }),
+    },
+    {
+      name: 'with shorthand (2-value)',
+      startSnippet: makeCodeSnippetWithKeyValue({ width: 120, height: 120, padding: 10 }),
+      before: async (renderResult: EditorRenderResult) => {
+        await act(async () => {
+          fireEvent.click(screen.getByTestId('padding-cycle-mode'))
+          await renderResult.getDispatchFollowUpActionsFinished()
+        })
+      },
+      control: async (renderResult: EditorRenderResult) => {
+        await expectSingleUndoStep(renderResult, async () => {
+          await setControlValue('padding-H', '20', renderResult.renderedDOM)
+        })
+      },
+      endSnippet: makeCodeSnippetWithKeyValue({ width: 140, height: 120, padding: '10px 20px' }),
+    },
+    {
+      name: 'single value with no other values set',
+      startSnippet: makeCodeSnippetWithKeyValue({ width: 100, height: 100 }),
+      before: async (renderResult: EditorRenderResult) => {
+        await act(async () => {
+          fireEvent.click(screen.getByTestId('padding-cycle-mode'))
+          await renderResult.getDispatchFollowUpActionsFinished()
+        })
+      },
+      control: async (renderResult: EditorRenderResult) => {
+        await expectSingleUndoStep(renderResult, async () => {
+          await setControlValue('padding-R', '20', renderResult.renderedDOM)
+        })
+      },
+      endSnippet: makeCodeSnippetWithKeyValue({ width: 120, height: 100, paddingRight: 20 }),
+    },
+    {
+      name: 'single value with shorthand set',
+      startSnippet: makeCodeSnippetWithKeyValue({ width: 120, height: 120, padding: 10 }),
+      before: async (renderResult: EditorRenderResult) => {
+        await act(async () => {
+          fireEvent.click(screen.getByTestId('padding-cycle-mode'))
+          await renderResult.getDispatchFollowUpActionsFinished()
+          fireEvent.click(screen.getByTestId('padding-cycle-mode'))
+          await renderResult.getDispatchFollowUpActionsFinished()
+        })
+      },
+      control: async (renderResult: EditorRenderResult) => {
+        await expectSingleUndoStep(renderResult, async () => {
+          await setControlValue('padding-R', '20', renderResult.renderedDOM)
+        })
+      },
+      endSnippet: makeCodeSnippetWithKeyValue({
+        width: 130,
+        height: 120,
+        padding: '10px 20px 10px 10px',
+      }),
+    },
+  ]
+
+  tests.forEach((tt, idx) => {
+    it(`(${idx + 1}) padding controls shorthand: ${tt.name}`, async () => {
+      const renderResult = await renderTestEditorWithCode(
+        makeTestProjectCodeWithSnippet(tt.startSnippet),
+        'await-first-dom-report',
+      )
+
+      const targetPath = EP.appendNewElementPath(TestScenePath, ['aaa', 'bbb'])
+
+      await act(async () => {
+        await renderResult.dispatch([selectComponents([targetPath], false)], false)
+      })
+
+      if (tt.before != null) {
+        await tt.before(renderResult)
+        await renderResult.getDispatchFollowUpActionsFinished()
+      }
+      await tt.control(renderResult)
+
+      await act(async () => {
+        const dispatchDone = renderResult.getDispatchFollowUpActionsFinished()
+        await renderResult.dispatch([selectComponents([targetPath], false)], true)
+        await dispatchDone
+      })
+
+      await renderResult.getDispatchFollowUpActionsFinished()
+
+      expect(getPrintedUiJsCode(renderResult.getEditorState())).toEqual(
+        makeTestProjectCodeWithSnippet(tt.endSnippet),
+      )
+    })
+  })
+})
+
+describe('canvas padding controls from the inspector', () => {
+  function makeCodeSnippetWithKeyValue(props: { [key: string]: any }): string {
+    const propsStr = Object.keys(props)
+      .map((k) => `${k}: ${JSON.stringify(props[k])},`)
+      .join('\n')
+    return `
+      <div
+        data-uid='aaa'
+      >
+        <div
+          style={{ ${propsStr} }}
+          data-uid='bbb'
+        >test</div>
+      </div>
+  `
+  }
+
+  const tests = [
+    {
+      name: 'single value shows controls on all sides',
+      startSnippet: makeCodeSnippetWithKeyValue({ padding: '50px' }),
+      controlTestID: 'padding-one',
+      hoveredCanvasControls: [
+        getSubduedPaddingControlTestID('top', 'hovered'),
+        getSubduedPaddingControlTestID('right', 'hovered'),
+        getSubduedPaddingControlTestID('bottom', 'hovered'),
+        getSubduedPaddingControlTestID('left', 'hovered'),
+      ],
+      focusedCanvasControls: [
+        getSubduedPaddingControlTestID('top', 'focused'),
+        getSubduedPaddingControlTestID('right', 'focused'),
+        getSubduedPaddingControlTestID('bottom', 'focused'),
+        getSubduedPaddingControlTestID('left', 'focused'),
+      ],
+    },
+    {
+      name: 'per-direction H value shows controls on horizontal sides',
+      startSnippet: makeCodeSnippetWithKeyValue({ padding: '50px 60px' }),
+      controlTestID: 'padding-H',
+      hoveredCanvasControls: [
+        getSubduedPaddingControlTestID('right', 'hovered'),
+        getSubduedPaddingControlTestID('left', 'hovered'),
+      ],
+      focusedCanvasControls: [
+        getSubduedPaddingControlTestID('right', 'focused'),
+        getSubduedPaddingControlTestID('left', 'focused'),
+      ],
+    },
+    {
+      name: 'per-direction V value shows controls on vertical sides',
+      startSnippet: makeCodeSnippetWithKeyValue({ padding: '50px 60px' }),
+      controlTestID: 'padding-V',
+      hoveredCanvasControls: [
+        getSubduedPaddingControlTestID('top', 'hovered'),
+        getSubduedPaddingControlTestID('bottom', 'hovered'),
+      ],
+      focusedCanvasControls: [
+        getSubduedPaddingControlTestID('top', 'focused'),
+        getSubduedPaddingControlTestID('bottom', 'focused'),
+      ],
+    },
+    {
+      name: 'per-side T value shows controls on top side',
+      startSnippet: makeCodeSnippetWithKeyValue({ padding: '50px 60px 70px 80px' }),
+      controlTestID: 'padding-T',
+      hoveredCanvasControls: [getSubduedPaddingControlTestID('top', 'hovered')],
+      focusedCanvasControls: [getSubduedPaddingControlTestID('top', 'focused')],
+    },
+    {
+      name: 'per-side R value shows controls on right side',
+      startSnippet: makeCodeSnippetWithKeyValue({ padding: '50px 60px 70px 80px' }),
+      controlTestID: 'padding-R',
+      hoveredCanvasControls: [getSubduedPaddingControlTestID('right', 'hovered')],
+      focusedCanvasControls: [getSubduedPaddingControlTestID('right', 'focused')],
+    },
+    {
+      name: 'per-side B value shows controls on bottom side',
+      startSnippet: makeCodeSnippetWithKeyValue({ padding: '50px 60px 70px 80px' }),
+      controlTestID: 'padding-B',
+      hoveredCanvasControls: [getSubduedPaddingControlTestID('bottom', 'hovered')],
+      focusedCanvasControls: [getSubduedPaddingControlTestID('bottom', 'focused')],
+    },
+    {
+      name: 'per-side L value shows controls on left side',
+      startSnippet: makeCodeSnippetWithKeyValue({ padding: '50px 60px 70px 80px' }),
+      controlTestID: 'padding-L',
+      hoveredCanvasControls: [getSubduedPaddingControlTestID('left', 'hovered')],
+      focusedCanvasControls: [getSubduedPaddingControlTestID('left', 'focused')],
+    },
+  ]
+
+  tests.forEach((t) => {
+    it(`${t.name} when hovering and focusing`, async () => {
+      const renderResult = await renderTestEditorWithCode(
+        makeTestProjectCodeWithSnippet(t.startSnippet),
+        'await-first-dom-report',
+      )
+
+      const targetPath = EP.appendNewElementPath(TestScenePath, ['aaa', 'bbb'])
+
+      await act(async () => {
+        await renderResult.dispatch([selectComponents([targetPath], false)], false)
+      })
+
+      const control = await getControl(t.controlTestID, renderResult.renderedDOM)
+
+      // Check the controls show when hovering
+      fireEvent.mouseEnter(control)
+      await renderResult.getDispatchFollowUpActionsFinished()
+
+      const hoveredControls = t.hoveredCanvasControls.flatMap((expectedControl) =>
+        renderResult.renderedDOM.queryAllByTestId(expectedControl),
+      )
+      expect(hoveredControls.length).toEqual(t.hoveredCanvasControls.length)
+
+      // Check the controls show when focusing
+      fireEvent.focus(control)
+      await renderResult.getDispatchFollowUpActionsFinished()
+
+      const focusedControls = t.focusedCanvasControls.flatMap((expectedControl) =>
+        renderResult.renderedDOM.queryAllByTestId(expectedControl),
+      )
+      expect(focusedControls.length).toEqual(t.focusedCanvasControls.length)
+    })
+  })
+})
